CS-36: hide close button until tap
This kills two birds with one stone: CS-24 requests that the close button hides until you tap.  The issue is the first tap is swallowed by the video view and then you have to tap again (CS-36).  This is kind of cheating, but it feels like a better UI to me.