This adds the ability to get Twilio stats using some code "borrowed" from CO
Unfortunately, the code is in Swift, and in order to call it from Obj-C, I have to include the "*-Swift.h" header, which HAS to be the Module name of the App being built, which the plugin doesn't know yet!  I have no idea how to solve this...but that's why the .m files have

#import "Spectator-Swift.h"

in them, which is totally the wrong thing to do.

Options:
1- Leave it like this and pretend it is ok
2- Rewrite that great Swift code in Obj-C
3- Give up and don't implement these stats